move get_batch_size to qml.math
